requirements.txt
absl-py==2.0.0
aiofiles==23.2.1
altair==5.1.2
altgraph==0.17.4
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
asttokens==2.4.0
attrs==23.1.0
backcall==0.2.0
bidict==0.22.1
blinker==1.6.2
bz2file==0.98
certifi==2023.7.22
cffi==1.16.0
chainer==7.8.1
charset-normalizer==3.3.0
ciseau==1.0.1
click==8.1.7
colorama==0.4.6
coloredlogs==15.0.1
comm==0.1.4
contourpy==1.1.1
controlnet-aux==0.0.3
cryptography==41.0.5
cycler==0.12.0
debugpy==1.8.0
decorator==5.1.1
diffusers==0.16.1
einops==0.7.0
epub==0.5.2
epub-conversion==1.0.15
exceptiongroup==1.1.3
executing==2.0.0
fastapi==0.103.2
ffmpy==0.3.1
filelock==3.12.4
Flask==2.2.3
Flask-Cors==4.0.0
Flask-SocketIO==5.3.6
flaskwebgui==0.3.5
flatbuffers==23.5.26
fonttools==4.43.1
fsspec==2023.9.2
gradio==3.47.1
gradio_client==0.6.0
h11==0.14.0
httpcore==0.18.0
httpx==0.25.0
huggingface-hub==0.17.3
humanfriendly==10.0
idna==3.4
imageio==2.31.5
importlib-metadata==6.8.0
importlib-resources==6.1.0
ipykernel==6.25.2
ipython==8.16.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter_client==8.3.1
jupyter_core==5.3.2
kiwisolver==1.4.5
lama-cleaner==1.2.4
lazy_loader==0.3
llvmlite==0.41.1
loguru==0.7.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.0
matplotlib-inline==0.1.6
mdurl==0.1.2
mediapipe==0.10.5
mpmath==1.3.0
mutagen==1.47.0
nest-asyncio==1.5.8
networkx==3.1
numba==0.58.1
numpy==1.26.0
omegaconf==2.3.0
onnxruntime==1.16.2
opencv-contrib-python==4.8.1.78
opencv-python==4.8.1.78
opencv-python-headless==4.8.1.78
orjson==3.9.7
packaging==23.1
pandas==2.1.1
parso==0.8.3
pefile==2023.2.7
pickleshare==0.7.5
piexif==1.1.3
Pillow==9.5.0
pillow-avif-plugin==1.4.1
platformdirs==3.11.0
pooch==1.8.0
prompt-toolkit==1.0.14
protobuf==3.20.3
psutil==5.9.5
pure-eval==0.2.2
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
pydub==0.25.1
pyftpdlib==1.5.9
Pygments==2.16.1
PyInquirer==1.0.3
pyinstaller==6.0.0
pyinstaller-hooks-contrib==2023.9
PyMatting==1.1.11
pyparsing==3.1.1
pyreadline3==3.4.1
python-dateutil==2.8.2
python-engineio==4.7.1
python-multipart==0.0.6
python-socketio==5.9.0
pytz==2023.3.post1
pywin32==306
pywin32-ctypes==0.2.2
PyYAML==6.0.1
pyzmq==25.1.1
referencing==0.30.2
regex==2023.10.3
rembg==2.0.52
requests==2.31.0
rich==13.6.0
rpds-py==0.10.4
safetensors==0.4.0
scikit-image==0.22.0
scipy==1.11.3
semantic-version==2.10.0
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.0
sounddevice==0.4.6
stack-data==0.6.3
starlette==0.27.0
sympy==1.12
tifffile==2023.9.26
timm==0.9.7
tokenizers==0.13.3
toolz==0.12.0
torch==2.1.0+cu118
torchaudio==2.1.0+cu118
torchvision==0.16.0
tornado==6.3.3
tqdm==4.66.1
traitlets==5.11.2
transformers==4.27.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.6
uvicorn==0.23.2
waifu2x==2023
Wand==0.6.13
wcwidth==0.2.8
websockets==11.0.3
Werkzeug==2.2.2
whichcraft==0.6.1
win32-setctime==1.1.0
wsproto==1.2.0
yacs==0.1.8
zipp==3.16.2
